+++ ietfparse/__init__.py
@@ -1,2 +1,2 @@
-version_info = (1, 2, 1)
+version_info = (1, 2, 2)
 __version__ = '.'.join(str(x) for x in version_info)